requirements-dev.txt
certifi==2019.3.9
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.1
django-cors-headers==2.5.3
django-crispy-forms==1.7.2
django-filter==2.1.0
django-markdown-deux==1.0.5
django-oauth-toolkit==1.2.0
django-pagedown==1.0.6
djangorestframework==3.9.3
djangorestframework-jwt==1.11.0
djangorestframework-oauth==1.1.0
idna==2.8
Markdown==3.1
markdown2==2.3.7
oauthlib==3.0.1
Pillow==6.0.0
PyJWT==1.7.1
python-decouple==3.1
pytz==2019.1
requests==2.21.0
sqlparse==0.3.0
static3==0.7.0
urllib3==1.24.3
